backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose
passlib[bcrypt]
alembic
psycopg2-binary  # if postgres
openai
python-dotenv

